requirements.txt
# 크롤링
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0

# 데이터 처리
pandas>=2.0.0
tqdm>=4.66.0

# LLM 전처리 (Gemini API)
google-generativeai>=0.3.0
